namespace refactoring and improvement on unit test
