+++ src/TeleScope.MSTest/TeleScope.MSTest.csproj
@@ -32,8 +32,8 @@
     <ProjectReference Include="..\TeleScope.Persistence.Json\TeleScope.Persistence.Json.csproj" />
     <ProjectReference Include="..\TeleScope.Persistence.Parquet\TeleScope.Persistence.Parquet.csproj" />
     <ProjectReference Include="..\TeleScope.Persistence.Yaml\TeleScope.Persistence.Yaml.csproj" />
-    <ProjectReference Include="..\TeleScope.UI.Abstractions\TeleScope.UI.Abstractions.csproj" />
     <ProjectReference Include="..\TeleScope.UI.Cli\TeleScope.UI.Cli.csproj" />
+    <ProjectReference Include="..\TeleScope.UI.Permissions\TeleScope.UI.Permissions.csproj" />
   </ItemGroup>
 
   <ItemGroup>
