+++ src/TeleScope.UI.Permissions/TeleScope.UI.Permissions.csproj
